+++ sdk/formrecognizer/azure-ai-formrecognizer/azure/ai/formrecognizer/_polling.py
@@ -80,7 +80,7 @@ class DocumentModelAdministrationLROPoller(Protocol[PollingReturnType]):
 
 class DocumentModelAdministrationClientLROPoller(LROPoller[PollingReturnType]):
     """Custom poller for model build operations. Call `result()` on the poller to return
-    a :class:`~azure.ai.formrecognizer.DocumentModelInfo`.
+    a :class:`~azure.ai.formrecognizer.DocumentModelDetails`.
 
     .. versionadded:: 2021-09-30-preview
         The *DocumentModelAdministrationLROPoller* poller object