+++ sdk/formrecognizer/azure-ai-formrecognizer/azure/ai/formrecognizer/aio/_document_model_administration_client_async.py
@@ -26,11 +26,11 @@ from .._api_versions import DocumentAnalysisApiVersion
 from .._polling import DocumentModelAdministrationPolling
 from .._models import (
     DocumentBuildMode,
-    DocumentModelInfo,
+    DocumentModelDetails,
     DocumentModelSummary,
-    ModelOperation,
-    ModelOperationInfo,
-    ResourceInfo,
+    ModelOperationDetails,
+    ModelOperationSummary,
+    ResourceDetails,
     TargetAuthorization,
 )
 
@@ -99,7 +99,7 @@ class DocumentModelAdministrationClient(FormRecognizerClientBaseAsync):
     @distributed_trace_async
     async def begin_build_model(
         self, source: str, build_mode: Union[str, DocumentBuildMode], **kwargs: Any
-    ) -> AsyncDocumentModelAdministrationLROPoller[DocumentModelInfo]:
+    ) -> AsyncDocumentModelAdministrationLROPoller[DocumentModelDetails]:
         """Build a custom model.
 
         The request must include a `source` parameter that is an
@@ -124,8 +124,8 @@ class DocumentModelAdministrationClient(FormRecognizerClientBaseAsync):
         :keyword tags: List of user defined key-value tag attributes associated with the model.
         :paramtype tags: dict[str, str]
         :return: An instance of an AsyncDocumentModelAdministrationLROPoller. Call `result()` on the poller
-            object to return a :class:`~azure.ai.formrecognizer.DocumentModelInfo`.
-        :rtype: ~azure.ai.formrecognizer.aio.AsyncDocumentModelAdministrationLROPoller[DocumentModelInfo]
+            object to return a :class:`~azure.ai.formrecognizer.DocumentModelDetails`.
+        :rtype: ~azure.ai.formrecognizer.aio.AsyncDocumentModelAdministrationLROPoller[DocumentModelDetails]
         :raises ~azure.core.exceptions.HttpResponseError:
 
         .. versionadded:: v2022-01-30-preview
@@ -148,7 +148,7 @@ class DocumentModelAdministrationClient(FormRecognizerClientBaseAsync):
             model_info = self._deserialize(
                 self._generated_models.ModelInfo, op_response.result
             )
-            return DocumentModelInfo._from_generated(model_info)
+            return DocumentModelDetails._from_generated(model_info)
 
         description = kwargs.pop("description", None)
         model_id = kwargs.pop("model_id", None)
@@ -186,7 +186,7 @@ class DocumentModelAdministrationClient(FormRecognizerClientBaseAsync):
     @distributed_trace_async
     async def begin_compose_model(
         self, component_model_ids: List[str], **kwargs: Any
-    ) -> AsyncDocumentModelAdministrationLROPoller[DocumentModelInfo]:
+    ) -> AsyncDocumentModelAdministrationLROPoller[DocumentModelDetails]:
         """Creates a composed model from a collection of existing models.
 
         A composed model allows multiple models to be called with a single model ID. When a document is
@@ -200,8 +200,8 @@ class DocumentModelAdministrationClient(FormRecognizerClientBaseAsync):
         :keyword tags: List of user defined key-value tag attributes associated with the model.
         :paramtype tags: dict[str, str]
         :return: An instance of an AsyncDocumentModelAdministrationLROPoller. Call `result()` on the poller
-            object to return a :class:`~azure.ai.formrecognizer.DocumentModelInfo`.
-        :rtype: ~azure.ai.formrecognizer.aio.AsyncDocumentModelAdministrationLROPoller[DocumentModelInfo]
+            object to return a :class:`~azure.ai.formrecognizer.DocumentModelDetails`.
+        :rtype: ~azure.ai.formrecognizer.aio.AsyncDocumentModelAdministrationLROPoller[DocumentModelDetails]
         :raises ~azure.core.exceptions.HttpResponseError:
 
         .. versionadded:: v2022-01-30-preview
@@ -226,7 +226,7 @@ class DocumentModelAdministrationClient(FormRecognizerClientBaseAsync):
             model_info = self._deserialize(
                 self._generated_models.ModelInfo, op_response.result
             )
-            return DocumentModelInfo._from_generated(model_info)
+            return DocumentModelDetails._from_generated(model_info)
 
         model_id = kwargs.pop("model_id", None)
         description = kwargs.pop("description", None)
@@ -300,7 +300,7 @@ class DocumentModelAdministrationClient(FormRecognizerClientBaseAsync):
     @distributed_trace_async
     async def begin_copy_model_to(
         self, model_id: str, target: TargetAuthorization, **kwargs: Any
-    ) -> AsyncDocumentModelAdministrationLROPoller[DocumentModelInfo]:
+    ) -> AsyncDocumentModelAdministrationLROPoller[DocumentModelDetails]:
         """Copy a model stored in this resource (the source) to the user specified
         target Form Recognizer resource.
 
@@ -313,8 +313,8 @@ class DocumentModelAdministrationClient(FormRecognizerClientBaseAsync):
             The copy authorization generated from the target resource's call to
             :func:`~get_copy_authorization()`.
         :return: An instance of a AsyncDocumentModelAdministrationLROPoller. Call `result()` on the poller
-            object to return a :class:`~azure.ai.formrecognizer.DocumentModelInfo`.
-        :rtype: ~azure.ai.formrecognizer.aio.AsyncDocumentModelAdministrationLROPoller[DocumentModelInfo]
+            object to return a :class:`~azure.ai.formrecognizer.DocumentModelDetails`.
+        :rtype: ~azure.ai.formrecognizer.aio.AsyncDocumentModelAdministrationLROPoller[DocumentModelDetails]
         :raises ~azure.core.exceptions.HttpResponseError:
 
         .. admonition:: Example:
@@ -334,7 +334,7 @@ class DocumentModelAdministrationClient(FormRecognizerClientBaseAsync):
             model_info = self._deserialize(
                 self._generated_models.ModelInfo, op_response.result
             )
-            return DocumentModelInfo._from_generated(model_info)
+            return DocumentModelDetails._from_generated(model_info)
 
         if not model_id:
             raise ValueError("model_id cannot be None or empty.")
@@ -418,33 +418,33 @@ class DocumentModelAdministrationClient(FormRecognizerClientBaseAsync):
         )
 
     @distributed_trace_async
-    async def get_resource_info(self, **kwargs: Any) -> ResourceInfo:
+    async def get_resource_details(self, **kwargs: Any) -> ResourceDetails:
         """Get information about the models under the Form Recognizer resource.
 
         :return: Summary of models under the resource - model count and limit.
-        :rtype: ~azure.ai.formrecognizer.ResourceInfo
+        :rtype: ~azure.ai.formrecognizer.ResourceDetails
         :raises ~azure.core.exceptions.HttpResponseError:
 
         .. admonition:: Example:
 
             .. literalinclude:: ../samples/v3.2-beta/async_samples/sample_manage_models_async.py
-                :start-after: [START get_resource_info_async]
-                :end-before: [END get_resource_info_async]
+                :start-after: [START get_resource_details_async]
+                :end-before: [END get_resource_details_async]
                 :language: python
                 :dedent: 4
                 :caption: Get model counts and limits under the Form Recognizer resource.
         """
 
         response = await self._client.get_info(**kwargs)
-        return ResourceInfo._from_generated(response.custom_document_models)
+        return ResourceDetails._from_generated(response.custom_document_models)
 
     @distributed_trace_async
-    async def get_model(self, model_id: str, **kwargs: Any) -> DocumentModelInfo:
+    async def get_model(self, model_id: str, **kwargs: Any) -> DocumentModelDetails:
         """Get a model by its ID.
 
         :param str model_id: Model identifier.
-        :return: DocumentModelInfo
-        :rtype: ~azure.ai.formrecognizer.DocumentModelInfo
+        :return: DocumentModelDetails
+        :rtype: ~azure.ai.formrecognizer.DocumentModelDetails
         :raises ~azure.core.exceptions.HttpResponseError or ~azure.core.exceptions.ResourceNotFoundError:
 
         .. admonition:: Example:
@@ -461,18 +461,18 @@ class DocumentModelAdministrationClient(FormRecognizerClientBaseAsync):
             raise ValueError("model_id cannot be None or empty.")
 
         response = await self._client.get_model(model_id=model_id, **kwargs)
-        return DocumentModelInfo._from_generated(response)
+        return DocumentModelDetails._from_generated(response)
 
     @distributed_trace
-    def list_operations(self, **kwargs: Any) -> AsyncItemPaged[ModelOperationInfo]:
+    def list_operations(self, **kwargs: Any) -> AsyncItemPaged[ModelOperationSummary]:
         """List information for each document model operation.
 
         Lists all document model operations associated with the Form Recognizer resource.
         Note that operation information only persists for 24 hours. If the operation was successful,
         the document model can be accessed using the :func:`~get_model` or :func:`~list_models` APIs.
 
-        :return: A pageable of ModelOperationInfo.
-        :rtype: ~azure.core.async_paging.AsyncItemPaged[ModelOperationInfo]
+        :return: A pageable of ModelOperationSummary.
+        :rtype: ~azure.core.async_paging.AsyncItemPaged[ModelOperationSummary]
         :raises ~azure.core.exceptions.HttpResponseError:
 
         .. admonition:: Example:
@@ -488,13 +488,13 @@ class DocumentModelAdministrationClient(FormRecognizerClientBaseAsync):
         return self._client.get_operations(  # type: ignore
             cls=kwargs.pop(
                 "cls",
-                lambda objs: [ModelOperationInfo._from_generated(x) for x in objs],
+                lambda objs: [ModelOperationSummary._from_generated(x) for x in objs],
             ),
             **kwargs
         )
 
     @distributed_trace_async
-    async def get_operation(self, operation_id: str, **kwargs: Any) -> ModelOperation:
+    async def get_operation(self, operation_id: str, **kwargs: Any) -> ModelOperationDetails:
         """Get a document model operation by its ID.
 
         Get a document model operation associated with the Form Recognizer resource.
@@ -502,8 +502,8 @@ class DocumentModelAdministrationClient(FormRecognizerClientBaseAsync):
         the document model can be accessed using the :func:`~get_model` or :func:`~list_models` APIs.
 
         :param str operation_id: The operation ID.
-        :return: ModelOperation
-        :rtype: ~azure.ai.formrecognizer.ModelOperation
+        :return: ModelOperationDetails
+        :rtype: ~azure.ai.formrecognizer.ModelOperationDetails
         :raises ~azure.core.exceptions.HttpResponseError:
 
         .. admonition:: Example:
@@ -519,7 +519,7 @@ class DocumentModelAdministrationClient(FormRecognizerClientBaseAsync):
         if not operation_id:
             raise ValueError("'operation_id' cannot be None or empty.")
 
-        return ModelOperation._from_generated(
+        return ModelOperationDetails._from_generated(
             await self._client.get_operation(operation_id, **kwargs),
             api_version=self._api_version,
         )